POSIX Semaphores, Shared Memory, No-failure voting process

+++ app/src/_app.c
@@ -10,40 +10,35 @@ pid_t* pids;            // PIDs array
 // -- Signals: Broadcast
 void __send_broadcast_signal(int signo) {
     for(int i = 0; i < processesCount; i++) {
-        kill(pids[i], signo);
+        if (pids[i] != -1) {
+            kill(pids[i], signo);
+        }
     }
 }
 
 // -- Signals: Handler
 void __master_sig_handler(int signo) {
     if(signo == SIGINT) {
-        printf("MASTER[%d]: Sending broadcast: %d\n", getpid(), signo);
         __send_broadcast_signal(SIGUSR1);
     } else if (signo == SIGUSR1) {
-        printf("MASTER[%d]: Sending broadcast: %d\n", getpid(), signo);
         __send_broadcast_signal(signo);
     } else if (signo == SIGUSR2) {
-        printf("MASTER[%d]: Sending broadcast: %d\n", getpid(), signo);
         __send_broadcast_signal(signo);
     }
 }
 
 // -- App: Main init
 int app_init(int argc, char** argv) {
-    // -- Args parse: processes count
+    // -- Args parse: processesCount
     if (argc == 1) {
         processesCount = THREADS_COUNT_DEFAULT;
     } else if (argc == 2) {
         processesCount = atoi(argv[1]);
         if (processesCount == 0) {
-            fprintf(stdout, "Failure!\n");
-            fprintf(stderr, "-- Error: Invalid argument!\n");
-            return -1;
+            return 3;
         }
     } else {
-        fprintf(stdout, "Failure!\n");
-        fprintf(stderr, "-- Error: Invalid argument!\n");
-        return -1;
+        return 3;
     }
 
     // -- Signals handler
@@ -52,10 +47,13 @@ int app_init(int argc, char** argv) {
     signal(SIGUSR2, __master_sig_handler);
 
     // -- Pipes
-    if (__pipes_init(processesCount)) {
-        fprintf(stdout, "Failure!\n");
-        fprintf(stderr, "-- Error: Pipes create!\n");
-        return -2;
+    if (pipes_init(processesCount)) {
+        return 1;
+    }
+
+    // -- Semaphores
+    if (sync_init(processesCount)) {
+        return 2;
     }
 
     return 0;
@@ -63,29 +61,58 @@ int app_init(int argc, char** argv) {
 
 // -- App: Children init
 int child_processes_init() {
-    // Threads dynamic array
+    // Children PIDs array
     pids = (pid_t*) malloc(processesCount * sizeof(pid_t));
-    // Threads create func result
-    int res = 0;
 
-    /* *** Threads create and start *** */
+    // PIDs array preconfigure
     int i = 0;
     for(; i < processesCount; i++) {
-        pids[i] = fork();
-        if (pids[i] == 0) { 
-            // Child process
-            component_main();
-            exit(0);
-        }
+        pids[i] = -1;
     }
 
+    /* *** Children create and start *** */
+    pid_t pid = -1;
+    for(i = 0; i < processesCount; i++) {
+        pid = fork();
+        if (pid == 0) { 
+            // Child process
+            component_main();
+        } else if (pid == -1) {
+            // fork(): error -- Abort
+            __send_broadcast_signal(SIGINT);
+            return 1;
+        } else {
+            // Parent process
+            pids[i] = pid;
+        }
+    }
+    
     return 0;
 }
 
 // -- App: Children wait
-int child_processes_wait() {
+void child_processes_wait() {
+    
+    sleep(2);
+    kill(getpid(), SIGUSR2);
+
+    /* ** Wait for children processes ** */
     for(int i = 0; i < processesCount; i++) {
         wait(NULL);
     }
+}
+
+int clean() {
+    /* ** Cleaning ** */
+    // -- Pipes shutdown
+    if (pipes_shutdown(processesCount) != 0) {
+        return 1;
+    }
+
+    // -- Semaphores shutdown
+    if (sync_shutdown(processesCount) != 0) {
+        return 2;
+    }
+
     return 0;
 }
